stoggle/backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.28
aiosqlite==0.20.0

# HTTP / async
httpx==0.27.0
aiohttp==3.9.3

# Data
pykrx==1.0.46
pandas==2.2.1
numpy==1.26.4

# NLP / LLM
openai==1.14.0
langchain==0.1.12
langchain-openai==0.1.1
konlpy==0.6.0
scikit-learn==1.4.1.post1

# Task queue (선택사항)
celery==5.3.6
redis==5.0.3

# Utilities
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
dart-fss==0.4.4
beautifulsoup4==4.12.3
